LayoutTests/http/wpt/webauthn/public-key-credential-create-failure-u2f.https-expected.txt

CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.
CONSOLE MESSAGE: User gesture is not detected. To use the platform authenticator, call 'navigator.credentials.create' within user activated events.

PASS PublicKeyCredential's [[create]] with malformed APDU payload in a mock hid authenticator.
PASS PublicKeyCredential's [[create]] with malformed U2F register response in a mock hid authenticator.
PASS PublicKeyCredential's [[create]] with register command error in a mock hid authenticator.
PASS PublicKeyCredential's [[create]] with bogus command error in a mock hid authenticator.
PASS PublicKeyCredential's [[create]] with first exclude credential matched in a mock hid authenticator.
PASS PublicKeyCredential's [[create]] with second exclude credential matched in a mock hid authenticator.
PASS PublicKeyCredential's [[create]] with first exclude credential matched in a mock hid authenticator. Test of user presence.
